# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API description for [Moira Alert API](https://moira.readthedocs.io/en/latest/overview.html).",
        "title": "Moira Alert",
        "contact": {
            "name": "Contact Moira Team",
            "email": "opensource@skbkontur.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "master"
    },
    "basePath": "/api",
    "paths": {
...
        "/trigger/{triggerID}/render": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Render trigger metrics plot",
                "operationId": "render-trigger-metrics",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "Trigger ID",
                        "name": "triggerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "t1",
                        "description": "Target metric name",
                        "name": "targetName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "default": "-1hour",
                        "description": "Start time for metrics retrieval",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "default": "now",
                        "description": "End time for metrics retrieval",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "UTC",
                        "description": "Timezone for rendering",
                        "name": "timezone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "light",
                        "description": "Plot theme",
                        "name": "theme",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "false",
                        "description": "Fetch real-time data",
                        "name": "realtime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rendered plot image successfully"
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorInternalServerExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "server error during request handling"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "api.ErrorInvalidRequestExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with the ID does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "api.ErrorNotFoundExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with ID '66741a8c-c2ba-4357-a2c9-ee78e0e7' does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
    },
...
}